mv jsma symb
